requirements.txt
fastapi
uvicorn
jinja2
openai
google-cloud-storage
python-multipart

pytest
pytest-mock
